Import sysutils/pkg-config:
Pkg-config utility retrieves metadata about the installation of
software packages.  This metadata includes version, compiler and
linker flags, and dependency information.

This is the OpenBSD version and it is inspired by, and intended to
operate as a drop-in replacement to, the program of the same name
available from freedesktop.org.